Transfered work to personal desktop
My version (recently updated) of R was not compatible with the older packages that I was running. It became impossible for me to run any of the code after I updated R on my desktop (as R takes a few days to upload compatible versions of the packages with the new version of R).

Since this was not working, I did the next few steps on the remote desktop, and then copied my work over to this desktop so that my version control could still be updated on GitHub.

These steps consited of creating one more visual, reformatting, linking the proper references, and finally making comments to further explain my work.